+++ Quotations Board Backend/DTOs/Bonds/YieldCurve.cs
@@ -6,7 +6,7 @@ public class YieldCurve
     public string BondUsed { get; set; } = string.Empty;
     public DateTime IssueDate { get; set; }
     public DateTime MaturityDate { get; set; }
-    public decimal Coupon { get; set; }
+    public double Coupon { get; set; }
     public Boolean BenchMarkFound { get; set; } = false;
     public double Tenure { get; set; }
